+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- New `ImageLayer` `repeat_x`/`y` fields are now parsed from map image layers.
+
 ## [0.14.0]
 ### Added
 - Added a new crate feature `world` to enable support for parsing `World` files.